+++ src/main/java/com/stek/webdash/service/repository/RequestHostRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.stek.webdash.model.RequestHost;
+import com.stek.webdash.model.domain.RequestHost;
 
 @Repository
-public interface RequestHostRepository extends JpaRepository<RequestHost, Integer> {
+public interface RequestHostRepository extends JpaRepository<RequestHost, Long> {
 
 	Optional<RequestHost> findByIpAddress(String ipAddress);
